+++ app/page.tsx
@@ -1,50 +1,12 @@
-"use client";
-
-import Image from "next/image";
-import { Inter } from "next/font/google";
-import styles from "./page.module.css";
+import Home from "./Home";
 import "./globals.css";
-import SPINNER from "../public/assets/imgs/loading-spin.svg";
-import SearchBar from "@/components/SearchBar/SearchBar";
-import { ChangeEvent, useState } from "react";
-import TableSection from "@/components/TableSection";
 
-const inter = Inter({ subsets: ["latin"] });
-
-const Home = () => {
-  const [searchText, setSearchText] = useState<string>("");
-  const [currentState, setCurrentState] = useState<string>("");
-  const handleOnChangeEvent = (e: ChangeEvent<HTMLInputElement>) => {
-    setSearchText(e.target.value);
-  };
-  const handleOnChangeTypeEvent = () => {
-    console.log("handleOnChangeTypeEvent ----");
-  };
-  console.log(process.env.NEXT_PUBLIC_API_KEY);
+const App = () => {
   return (
-    <main className={styles.container}>
-      <h1>Holidays across the world</h1>
-      <div className={`${styles.flexCenter} ${styles.marginTop}`}>
-        <SearchBar
-          id={"search-box"}
-          type={"text"}
-          name={"search-box"}
-          value={searchText}
-          placeHolder={"Search country"}
-          currentState={currentState}
-          onChangeEvent={handleOnChangeEvent}
-          onChangeTypeEvent={handleOnChangeTypeEvent}
-        />
-        <div>{searchText}</div>
-      </div>
-      <TableSection tableData={{}} currentState={currentState} />
-      {currentState === "Loading" && (
-        <div className={styles.loading_spinner}>
-          <Image src={SPINNER} alt={"loading_spinner"} width={72} height={72} />
-        </div>
-      )}
-    </main>
+    <>
+      <Home />
+    </>
   );
 };
 
-export default Home;
+export default App;
